students/k3339/laboratory_works/Kotovshchikov_Andrey/laboratory_work_3/pyproject.toml
[tool.poetry]
name = "laboratory-work-3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
django-filter = "^24.3"
djoser = "^2.3.1"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
